# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Roadmap for Jira API",
        "description": "Endpoint mapping for Roadmap for Jira App.",
        "contact": {
            "email": "contact@mrrafael.ca"
        },
        "version": "2.0"
    },
    "servers": [
        {
            "url": "http://localhost:3030",
            "description": "Server"
        }
    ],
    "paths": {
...
        "/api/v1/issue/find": {
            "put": {
                "tags": [
                    "Issue"
                ],
                "description": "Load a List of Issues objects for current user.",
                "operationId": "findIssues",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FindIssueModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Issue list loaded successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing a list of Issues",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/IssueModelList"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the error message",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorModel"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/dash": {
            "put": {
                "tags": [
                    "Roadmap"
                ],
                "description": "Load a List of Roadmaps objects for current user.",
                "operationId": "getDash",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DashParamsModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Roadmap list loaded successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing a list of Roadmaps",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/RoadmapModelList"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the error message",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorModel"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/roadmap/": {
            "post": {
                "tags": [
                    "Roadmap"
                ],
                "description": "Create or update a Roadmap object.",
                "operationId": "SaveRoadmap",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoadmapModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Roadmap created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the saved Roadmap",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/RoadmapModel"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the error message",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorModel"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/roadmap/{guid}/{itens}": {
            "get": {
                "tags": [
                    "Roadmap"
                ],
                "description": "Load a Roadmap object by GUID.",
                "operationId": "LoadRoadmap",
                "parameters": [
                    {
                        "name": "guid",
                        "in": "path",
                        "description": "GUID of the Roadmap",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "6c71d92e-1af3-4597-a51c-e7689334b7ab"
                        }
                    },
                    {
                        "name": "itens",
                        "in": "path",
                        "description": "Load the Roadmap with the list of Issues (0-false/1-true).",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Roadmap created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the Roadmap identified by the GUID",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/RoadmapModel"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the error message",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorModel"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/roadmap/{guid}/sync": {
            "put": {
                "tags": [
                    "Roadmap"
                ],
                "description": "Synchronize Roadmap Issues with Jira (send Start/End date, get Issue details).",
                "operationId": "syncRoadmapIssues",
                "parameters": [
                    {
                        "name": "guid",
                        "in": "path",
                        "description": "GUID of the Roadmap",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "6c71d92e-1af3-4597-a51c-e7689334b7ab"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Roadmap created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the Roadmap identified by the GUID",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/RoadmapModel"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "JSON object containing the error message",
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ErrorModel"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FindIssueModel": {
                "required": [
                    "guid",
                    "filter",
                    "subTasks"
                ],
                "properties": {
                    "guid": {
                        "type": "string",
                        "example": "455bbfd4-abcd-1234-8345-2d29d164699b"
                    },
                    "filter": {
                        "type": "string",
                        "example": "project in (TES, ABC) or Key = XYZ-123"
                    },
                    "subTasks": {
                        "type": "boolean",
                        "example": "true"
                    }
                },
                "type": "object"
            },
...
            "IssueModel": {
                "required": [
                    "keyJira",
                    "summary",
                    "status",
                    "assignee",
                    "startDate",
                    "endDate",
                    "issueType",
                    "icoUrl",
                    "issueUrl",
                    "left",
                    "top",
                    "width",
                    "css",
                    "exists"
                ],
                "properties": {
                    "keyJira": {
                        "type": "string",
                        "example": "ABC-123"
                    },
                    "summary": {
                        "type": "string",
                        "example": "Check the new feature"
                    },
                    "status": {
                        "type": "string",
                        "example": "To Do"
                    },
                    "assignee": {
                        "type": "string",
                        "example": "John Wick"
                    },
                    "startDate": {
                        "type": "string",
                        "example": "2024-01-15"
                    },
                    "endDate": {
                        "type": "string",
                        "example": "2024-01-20"
                    },
                    "issueType": {
                        "type": "string",
                        "example": "Task"
                    },
                    "icoUrl": {
                        "type": "string",
                        "example": "https://jira.com/ico.png"
                    },
                    "issueUrl": {
                        "type": "string",
                        "example": "https://jira.com/ABC-123"
                    },
                    "left": {
                        "type": "number",
                        "example": "175"
                    },
                    "top": {
                        "type": "number",
                        "example": "250"
                    },
                    "width": {
                        "type": "number",
                        "example": "145"
                    },
                    "css": {
                        "type": "string",
                        "example": "done"
                    },
                    "exists": {
                        "type": "boolean",
                        "example": "true"
                    }
                },
                "type": "object"
            },
            "IssueModelList": {
                "required": [
                    "list"
                ],
                "properties": {
                    "list": {
                        "type": "array",
                        "items": {
                            "oneOf": [
                                {
                                    "$ref": "#/components/schemas/IssueModel"
                                }
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
            "RoadmapModel": {
                "required": [
                    "guid",
                    "name",
                    "description",
                    "start_date",
                    "end_date",
                    "active",
                    "shares"
                ],
                "properties": {
                    "guid": {
                        "type": "string",
                        "example": "6c71d92e-1af3-4597-a51c-e7689334b7ab"
                    },
                    "name": {
                        "type": "string",
                        "example": "July 2024 - Roadmap"
                    },
                    "description": {
                        "type": "string",
                        "example": "This is a test description"
                    },
                    "start_date": {
                        "type": "string",
                        "example": "2024-07-01"
                    },
                    "end_date": {
                        "type": "string",
                        "example": "2024-07-31"
                    },
                    "active": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "level": {
                        "description": "0 - Owner, 1 - Editor, 2 - Viewer",
                        "type": "integer",
                        "example": "0"
                    },
                    "issues": {
                        "type": "array",
                        "items": {
                            "oneOf": [
                                {
                                    "$ref": "#/components/schemas/IssueModel"
                                }
                            ]
                        }
                    },
                    "shares": {
                        "type": "array",
                        "items": {
                            "oneOf": [
                                {
                                    "$ref": "#/components/schemas/ShareModel"
                                }
                            ]
                        }
                    }
                },
                "type": "object"
            },
            "RoadmapModelList": {
                "properties": {
                    "list": {
                        "type": "array",
                        "items": {
                            "oneOf": [
                                {
                                    "$ref": "#/components/schemas/RoadmapModel"
                                }
                            ]
                        },
                        "example": [
                            {
                                "guid": "6c71d92e-1af3-4597-a51c-e7689334b7ab",
                                "name": "July 2024 - Roadmap",
                                "description": "This is a test description",
                                "start_date": "2024-07-01",
                                "end_date": "2024-07-31",
                                "active": true
                            },
                            {
                                "guid": "6c71d92e-1af3-4597-a51c-e7689334b7ab",
                                "name": "July 2024 - Roadmap",
                                "description": "This is a test description",
                                "start_date": "2024-07-01",
                                "end_date": "2024-07-31",
                                "active": true
                            }
                        ]
                    }
                },
                "type": "object"
            },
            "DashParamsModel": {
                "required": [
                    "filter"
                ],
                "properties": {
                    "filter": {
                        "required": [
                            "active"
                        ],
                        "properties": {
                            "active": {
                                "description": "0 - Inactive, 1 - Active, 2 - All",
                                "type": "integer",
                                "example": "0"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "ShareModel": {
                "required": [
                    "id",
                    "read_only",
                    "user_email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "3"
                    },
                    "read_only": {
                        "type": "integer",
                        "example": "0"
                    },
                    "user_email": {
                        "type": "string",
                        "example": "john@test.com"
                    }
                },
                "type": "object"
            },
...
            "ErrorModel": {
                "required": [
                    "code",
                    "error"
                ],
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "42S22"
                    },
                    "error": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "message": {
                        "type": "string",
                        "example": "Complementary Error message"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
}